Restructure files and Refactor start for removal of git submodule by adding NuGet package for Newtonsoft.Json.
