+++ Console.Calculator.cyanitol.sln
@@ -1,17 +1,12 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.002.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Academy.Console.Calculator", "src\Console.Calculator.cyanitol.csproj", "{7EC1AD83-A854-464F-A39C-A18EE01F9EF6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Console.Calculator.cyanitol", "Console.Calculator.cyanitol\Console.Calculator.cyanitol.csproj", "{02234CF4-235D-41D4-82D7-87ED7E895E6A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorLibrary", "Calculator\CalculatorLibrary.csproj", "{D6A035F0-9EF7-4EB4-A12A-4217AD6880C5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Menu", "Menu\Menu.csproj", "{4511521F-AE50-43CA-95C2-85079DF36811}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MenuLibrary", "Menu\MenuLibrary.csproj", "{8108A157-1C69-4EBF-B3D2-CFF832AA54F7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Logger", "Logger\Logger.csproj", "{CC506C3C-988B-4BF1-B596-B219C99F4F1B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LoggerLibrary", "Logger\LoggerLibrary.csproj", "{02ED16A9-FC9F-4598-9C30-921A31709299}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Newtonsoft.Json", "contrib\Newtonsoft.Json\Src\Newtonsoft.Json\Newtonsoft.Json.csproj", "{A97A4CB3-51AD-4605-BA9F-D6861E582C85}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator", "Calculator\Calculator.csproj", "{8CD32F9A-D9F2-4448-9814-743D18E92C90}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,27 +14,21 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7EC1AD83-A854-464F-A39C-A18EE01F9EF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7EC1AD83-A854-464F-A39C-A18EE01F9EF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7EC1AD83-A854-464F-A39C-A18EE01F9EF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7EC1AD83-A854-464F-A39C-A18EE01F9EF6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D6A035F0-9EF7-4EB4-A12A-4217AD6880C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D6A035F0-9EF7-4EB4-A12A-4217AD6880C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D6A035F0-9EF7-4EB4-A12A-4217AD6880C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D6A035F0-9EF7-4EB4-A12A-4217AD6880C5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8108A157-1C69-4EBF-B3D2-CFF832AA54F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8108A157-1C69-4EBF-B3D2-CFF832AA54F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8108A157-1C69-4EBF-B3D2-CFF832AA54F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8108A157-1C69-4EBF-B3D2-CFF832AA54F7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02ED16A9-FC9F-4598-9C30-921A31709299}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02ED16A9-FC9F-4598-9C30-921A31709299}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02ED16A9-FC9F-4598-9C30-921A31709299}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02ED16A9-FC9F-4598-9C30-921A31709299}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {957710A6-7730-4EC5-928B-54A3EE6D5153}
+		{02234CF4-235D-41D4-82D7-87ED7E895E6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02234CF4-235D-41D4-82D7-87ED7E895E6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02234CF4-235D-41D4-82D7-87ED7E895E6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02234CF4-235D-41D4-82D7-87ED7E895E6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4511521F-AE50-43CA-95C2-85079DF36811}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4511521F-AE50-43CA-95C2-85079DF36811}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4511521F-AE50-43CA-95C2-85079DF36811}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4511521F-AE50-43CA-95C2-85079DF36811}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CC506C3C-988B-4BF1-B596-B219C99F4F1B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CC506C3C-988B-4BF1-B596-B219C99F4F1B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CC506C3C-988B-4BF1-B596-B219C99F4F1B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CC506C3C-988B-4BF1-B596-B219C99F4F1B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8CD32F9A-D9F2-4448-9814-743D18E92C90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8CD32F9A-D9F2-4448-9814-743D18E92C90}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8CD32F9A-D9F2-4448-9814-743D18E92C90}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8CD32F9A-D9F2-4448-9814-743D18E92C90}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
