--- README.md
@@ -1,13 +1,0 @@
-# Academy.Console.Calculator
-
-## Complete Tutorial
-* Complete the following tutorial: https://docs.microsoft.com/en-us/visualstudio/get-started/csharp/tutorial-console?view=vs-2022 (parts 1 and 2): Create a Calculator App (Microsoft Docs). You can find a link to part two at the bottom of the page.
-
-## Challenges
-* Create a functionality that will count the amount of times the calculator was used.
-* Store a list with the latest calculations. And give the users the ability to delete that list.
-* Allow the users to use the results in the list above to perform new calculations.
-* Add extra calculations: Square Root, Taking the Power, 10x, Trigonometry functions.
-
-## Notes
-Json library added as a git submodule.  Remember to recurse on clone.
